src/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
web3
python-dotenv
poseidon-hash
